Added null check for StartOptions
StartOptions, AppBuilderConfigurator, and HttpConfigurator can all be
set to null, but are required to not be null. The exceptions thrown are
the same type that would be thrown anyway, but with additional messages
attached for troubleshooting.

This style was chosen for the following reasons:
1. Error messages should contain useful information where possible.
2. It is more efficient to wait until the method is potentially invoked,
since the check will only happen once.

However, there are good arguments for doing it other ways:
The null check could be performed at every assignment to these fields. This would allow a debugger to pinpoint exactly where the issue is at the cost of potentially being invoked multiple times (though there is no reason I can think of for doing this.)

The null checks could result in the default behaviors being assigned. However, silently assigning a different behavior than the client is expecting rather than
throwing an error seems like bad design.

The null checks could be omitted entirely.